- weapon system up and running

+++ code/branches/FICN/src/orxonox/objects/weapon_system/barrel_gun.h
@@ -48,11 +48,11 @@ namespace orxonox {
   protected:
     void primaryFire();
 
-    void primaryFiring(unsigned int);
+    void primaryFiring(float);
 
     void secondaryFire();
 
-    void secondaryFiring(unsigned int);
+    void secondaryFiring(float);
 
   public:
